Completely removed Slice from the project
It seemed like a good idea when I created it. But after I started storing separate audio channels together, it no longer had any real purpose. What it did do however, was make interop with libraries that implement the FFT harder than it needed to be. Which is why it has to go. Especially now that Span<T> is all that is needed.